refactor(session): delete event log store
