+++ packages/session/test/projection.test.ts
@@ -3,7 +3,6 @@ import {
 	applySnapshot,
 	emptyProjection,
 	hydrateDashboardProjection,
-	rebuildProjectionFromEventLog,
 	reduceProjectableEvent,
 	serializeDashboardProjection,
 } from "../src/projection.js";
@@ -243,43 +242,4 @@ test("snapshot clears stale current run ids", () => {
 	expect(repaired.work.get("work-1")?.status).toBe("done");
 	expect(repaired.work.get("work-1")?.currentRunId).toBeUndefined();
 	expect(repaired.attempts.size).toBe(0);
-});
-
-test("session_started starts a fresh live projection window", () => {
-	const projection = rebuildProjectionFromEventLog(
-		[
-			{
-				kind: "session_event",
-				sessionId: "session-1",
-				sequence: 1,
-				timestamp: "2026-06-29T10:00:00.000Z",
-				type: "attempt_started",
-				payload: { run: { runId: "run-0", workKey: "work-1" } },
-			},
-			{
-				kind: "agent_event",
-				sessionId: "session-1",
-				sequence: 2,
-				timestamp: "2026-06-29T10:00:01.000Z",
-				runId: "run-0",
-				event: {
-					type: "turn_end",
-					message: { usage: { totalTokens: 123, cost: { total: 0.45 } } },
-				},
-			},
-			{
-				kind: "session_event",
-				sessionId: "session-1",
-				sequence: 3,
-				timestamp: "2026-06-29T10:01:00.000Z",
-				type: "session_started",
-			},
-		],
-		emptyProjection("session-1", "workflow"),
-	);
-
-	expect(projection.usageTotals).toEqual({ tokens: 0 });
-	expect(projection.attempts.size).toBe(0);
-	expect(projection.work.size).toBe(0);
-	expect(projection.status).toBe("running");
 });
